+++ src/setImmutable.spec.ts
@@ -238,4 +238,18 @@ describe("setImmutable", () => {
             },
         });
     });
+
+    it("throws on prototype-polluting keys and keeps Object.prototype clean", () => {
+        (() => {
+            setImmutable({}, "__proto__.polluted", "yes");
+        }).must.throw(TypeError, /prototype-polluting/u);
+        (() => {
+            setImmutable({}, "constructor.prototype.polluted", "yes");
+        }).must.throw(TypeError, /prototype-polluting/u);
+        (() => {
+            setImmutable({}, ["a", "__proto__", "z"], "yes");
+        }).must.throw(TypeError, /prototype-polluting/u);
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        (({} as any).polluted === undefined).must.be.true();
+    });
 });