+++ CHANGELOG.md
@@ -3,7 +3,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.24.1+]
+THIS LIBRARY IS DEPRECATED. PLEASE SWITCH TO THE SUCCESSOR: `@ezez/utils`.
+
+## [0.24.3] - 2026-08-16
+### Fixed
+- possible prototype pollution on untrusted input on `set`, `setImmutable`, `merge`
+
+## [0.24.2] - 2023-03-10
 ### Dev
 - added missing docs
 - fixed some jsdoc strings (so they contain types for non-ts users)